+++ libraries/ACD10/CHANGELOG.md
@@ -6,8 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-## [0.1.0] - 2023-09-25
+## [0.1.1] - 2023-10-05
+- refactor **begin()**
+- add **ACD10_readSensor_wire1.ino** ESP32 et al.
+- update readme.md
+- minor edits
 
+
+## [0.1.0] - 2023-09-25
 - initial version